feat : company, jobposting Guards 연결 및 유효성 검사 추가

+++ src/jobposting/jobposting.service.ts
@@ -10,13 +10,34 @@ export class JobpostingService {
   constructor(
     @InjectRepository(Jobposting)
     private readonly jobpostingRepository: Repository<Jobposting>,
+    @InjectRepository(Company)
+    private readonly companyRepository: Repository<Company>,
   ) {}
 
   // 채용 공고 생성
   async createJobposting(
+    id: number,
     companyId: number,
     createJobpostingDto: CreateJobpostingDto,
   ): Promise<Jobposting> {
+    // 회사에 아이디가 생성되있는지 찾기
+    const existingCompany = await this.companyRepository.findOne({
+      where: { id: companyId },
+    });
+
+    // 데이터베이스에 회사가 생성되어있는지 확인
+    if (!existingCompany) {
+      throw new HttpException(
+        '회사를 찾을 수 없습니다.',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // 로그인된 회사 ID와 채용공고의 회사 ID 비교
+    if (companyId !== id) {
+      throw new HttpException('권한이 없습니다.', HttpStatus.FORBIDDEN);
+    }
+
     const {
       title,
       career,
@@ -48,20 +69,56 @@ export class JobpostingService {
 
   // 채용공고 전체 조회
   async findAllJobposting(companyId: number): Promise<Jobposting[]> {
+    const existingCompany = await this.jobpostingRepository.findOne({
+      where: { companyId },
+    });
+
+    if (!existingCompany) {
+      throw new HttpException(
+        '회사를 찾을 수 없습니다.',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return await this.jobpostingRepository.find({ where: { companyId } });
   }
 
   // 채용공고 1개 조회
-  async findOneJobposting(companyId: number, id: number): Promise<Jobposting> {
+  async findOneJobposting(
+    companyId: number,
+    jobpostingId: number,
+  ): Promise<Jobposting> {
+    // 회사 아이디가 있는지 확인
+    const existingCompany = await this.jobpostingRepository.findOne({
+      where: { companyId },
+    });
+
+    // 해당 id가 있는 채용 공고 있는지 확인
+    const existingJobposting = await this.jobpostingRepository.findOne({
+      where: { id: jobpostingId },
+    });
+
+    // 회사와 채용공고 id가 없으면 예외
+    if (!existingCompany || !existingJobposting) {
+      throw new HttpException(
+        '채용 공고를 찾을 수 없습니다.',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return await this.jobpostingRepository.findOne({
-      where: { companyId, id },
+      where: { companyId, id: jobpostingId },
     });
   }
 
   // 채용공고 수정
-  async updateJobposting(id: number, updateJobpostingDto: UpdateJobpostingDto) {
+  async updateJobposting(
+    jobpostingId: number,
+    id: number,
+    updateJobpostingDto: UpdateJobpostingDto,
+  ) {
     const jobposting = await this.jobpostingRepository.findOne({
-      where: { id },
+      where: { id: jobpostingId },
     });
 
     if (!jobposting) {
@@ -70,20 +127,33 @@ export class JobpostingService {
         HttpStatus.BAD_REQUEST,
       );
     }
-    Object.assign(jobposting, updateJobpostingDto);
+
+    console.log(jobposting.companyId);
+    console.log(id);
+    // 로그인된 회사 ID와 채용공고의 회사 ID 비교
+    if (jobposting.companyId !== id) {
+      throw new HttpException('권한이 없습니다.', HttpStatus.FORBIDDEN);
+    }
+
+    Object.assign(jobposting, id, updateJobpostingDto);
     return await this.jobpostingRepository.save(jobposting);
   }
 
   // 채용공고 삭제
-  async removeJobposting(id: number) {
+  async removeJobposting(jobpostingId: number, id: number) {
     const jobposting = await this.jobpostingRepository.findOne({
-      where: { id },
+      where: { id: jobpostingId },
     });
     if (!jobposting) {
       throw new HttpException(
         '채용 공고를 찾을 수 없습니다.',
         HttpStatus.BAD_REQUEST,
       );
+    }
+
+    // 로그인된 회사 ID와 채용공고의 회사 ID 비교
+    if (jobposting.companyId !== id) {
+      throw new HttpException('권한이 없습니다.', HttpStatus.FORBIDDEN);
     }
     return await this.jobpostingRepository.remove(jobposting);
   }
